chore: delete duplicated node args types
