--- bin/reth/src/args/database_args.rs
@@ -1,33 +1,0 @@
-//! clap [Args](clap::Args) for database configuration
-
-use clap::Args;
-use reth_interfaces::db::LogLevel;
-
-/// Parameters for database configuration
-#[derive(Debug, Args, PartialEq, Default, Clone, Copy)]
-#[clap(next_help_heading = "Database")]
-pub struct DatabaseArgs {
-    /// Database logging level. Levels higher than "notice" require a debug build.
-    #[arg(long = "db.log-level", value_enum)]
-    pub log_level: Option<LogLevel>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-
-    /// A helper type to parse Args more easily
-    #[derive(Parser)]
-    struct CommandParser<T: Args> {
-        #[clap(flatten)]
-        args: T,
-    }
-
-    #[test]
-    fn test_parse_database_args() {
-        let default_args = DatabaseArgs::default();
-        let args = CommandParser::<DatabaseArgs>::parse_from(["reth"]).args;
-        assert_eq!(args, default_args);
-    }
-}
