+++ bin/reth/src/commands/db/diff.rs
@@ -1,19 +1,9 @@
-use std::{
-    collections::HashMap,
-    fmt::Debug,
-    fs::{self, File},
-    hash::Hash,
-    io::Write,
-    path::{Path, PathBuf},
-};
-
-use crate::utils::DbTool;
-use clap::Parser;
-
-use crate::{
+use crate::node_core::{
     args::DatabaseArgs,
     dirs::{DataDirPath, PlatformPath},
+    utils::DbTool,
 };
+use clap::Parser;
 use reth_db::{
     cursor::DbCursorRO, database::Database, open_db_read_only, table::Table, transaction::DbTx,
     AccountChangeSet, AccountHistory, AccountsTrie, BlockBodyIndices, BlockOmmers,
@@ -21,6 +11,14 @@ use reth_db::{
     HeaderNumbers, HeaderTD, Headers, PlainAccountState, PlainStorageState, PruneCheckpoints,
     Receipts, StorageChangeSet, StorageHistory, StoragesTrie, SyncStage, SyncStageProgress, Tables,
     TransactionBlock, Transactions, TxHashNumber, TxSenders,
+};
+use std::{
+    collections::HashMap,
+    fmt::Debug,
+    fs::{self, File},
+    hash::Hash,
+    io::Write,
+    path::{Path, PathBuf},
 };
 use tracing::info;
