+++ bin/reth/src/commands/debug_cmd/build_block.rs
@@ -1,6 +1,15 @@
 //! Command for debugging block building.
 
-use crate::runner::CliContext;
+use crate::{
+    node_core::{
+        args::{
+            utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
+            DatabaseArgs,
+        },
+        dirs::{DataDirPath, MaybePlatformPath},
+    },
+    runner::CliContext,
+};
 use alloy_rlp::Decodable;
 use clap::Parser;
 use eyre::Context;
@@ -15,6 +24,8 @@ use reth_db::{init_db, DatabaseEnv};
 use reth_interfaces::{consensus::Consensus, RethResult};
 use reth_node_api::PayloadBuilderAttributes;
 use reth_payload_builder::database::CachedReads;
+#[cfg(not(feature = "optimism"))]
+use reth_payload_builder::EthPayloadBuilderAttributes;
 #[cfg(feature = "optimism")]
 use reth_payload_builder::OptimismPayloadBuilderAttributes;
 use reth_primitives::{
@@ -39,16 +50,6 @@ use reth_transaction_pool::{
 };
 use std::{path::PathBuf, str::FromStr, sync::Arc};
 use tracing::*;
-
-use crate::{
-    args::{
-        utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
-        DatabaseArgs,
-    },
-    dirs::{DataDirPath, MaybePlatformPath},
-};
-#[cfg(not(feature = "optimism"))]
-use reth_payload_builder::EthPayloadBuilderAttributes;
 
 /// `reth debug build-block` command
 /// This debug routine requires that the node is positioned at the block before the target.
