+++ bin/reth/src/commands/debug_cmd/in_memory_merkle.rs
@@ -1,14 +1,16 @@
 //! Command for debugging in-memory merkle trie calculation.
 
 use crate::{
-    args::{
-        get_secret_key,
-        utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
-        DatabaseArgs, NetworkArgs,
+    node_core::{
+        args::{
+            get_secret_key,
+            utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
+            DatabaseArgs, NetworkArgs,
+        },
+        dirs::{DataDirPath, MaybePlatformPath},
+        utils::{get_single_body, get_single_header},
     },
-    dirs::{DataDirPath, MaybePlatformPath},
     runner::CliContext,
-    utils::{get_single_body, get_single_header},
 };
 use backon::{ConstantBuilder, Retryable};
 use clap::Parser;
