+++ bin/reth/src/commands/debug_cmd/replay_engine.rs
@@ -1,10 +1,12 @@
 use crate::{
-    args::{
-        get_secret_key,
-        utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
-        DatabaseArgs, NetworkArgs,
+    node_core::{
+        args::{
+            get_secret_key,
+            utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
+            DatabaseArgs, NetworkArgs,
+        },
+        dirs::{DataDirPath, MaybePlatformPath},
     },
-    dirs::{DataDirPath, MaybePlatformPath},
     runner::CliContext,
 };
 use clap::Parser;
