+++ bin/reth/src/commands/recover/storage_tries.rs
@@ -1,7 +1,9 @@
 use crate::{
-    args::utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
-    dirs::{DataDirPath, MaybePlatformPath},
-    init::init_genesis,
+    node_core::{
+        args::utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
+        dirs::{DataDirPath, MaybePlatformPath},
+        init::init_genesis,
+    },
     runner::CliContext,
 };
 use clap::Parser;
