+++ bin/reth/src/commands/stage/dump/execution.rs
@@ -1,5 +1,5 @@
 use super::setup;
-use crate::utils::DbTool;
+use crate::node_core::utils::DbTool;
 use eyre::Result;
 use reth_db::{
     cursor::DbCursorRO, database::Database, table::TableImporter, tables, transaction::DbTx,
