+++ bin/reth/src/commands/stage/run.rs
@@ -3,14 +3,16 @@
 //! Stage debugging tool
 
 use crate::{
-    args::{
-        get_secret_key,
-        utils::{chain_help, chain_spec_value_parser, SUPPORTED_CHAINS},
-        DatabaseArgs, NetworkArgs, StageEnum,
+    node_core::{
+        args::{
+            get_secret_key,
+            utils::{chain_help, chain_spec_value_parser, SUPPORTED_CHAINS},
+            DatabaseArgs, NetworkArgs, StageEnum,
+        },
+        dirs::{DataDirPath, MaybePlatformPath},
+        version::SHORT_VERSION,
     },
-    dirs::{DataDirPath, MaybePlatformPath},
     prometheus_exporter,
-    version::SHORT_VERSION,
 };
 use clap::Parser;
 use reth_beacon_consensus::BeaconConsensus;
