+++ crates/consensus/auto-seal/tests/it/auto_mine.rs
@@ -2,22 +2,20 @@
 
 use clap::Parser;
 use jsonrpsee::{core::client::ClientT, http_client::HttpClientBuilder, rpc_params};
-
+use reth::{
+    commands::node::NodeCommand,
+    node_core::cli::{
+        components::RethNodeComponents,
+        ext::{NoArgs, NoArgsCliExt, RethNodeCommandConfig},
+    },
+    runner::CliRunner,
+    tasks::TaskSpawner,
+};
 use reth_primitives::{hex, revm_primitives::FixedBytes, ChainSpec, Genesis};
 use reth_provider::CanonStateSubscriptions;
 use reth_transaction_pool::TransactionPool;
 use std::{sync::Arc, time::Duration};
 use tokio::time::timeout;
-
-use reth::{
-    cli::{
-        components::RethNodeComponents,
-        ext::{NoArgs, NoArgsCliExt, RethNodeCommandConfig},
-    },
-    commands::node::NodeCommand,
-    runner::CliRunner,
-    tasks::TaskSpawner,
-};
 
 #[derive(Debug)]
 struct AutoMineConfig;
