+++ crates/node-core/src/args/network_args.rs
@@ -122,6 +122,21 @@ impl NetworkArgs {
 
         Some(peers_file)
     }
+
+    /// Sets the p2p port to zero, to allow the OS to assign a random unused port when
+    /// the network components bind to a socket.
+    pub fn with_unused_p2p_port(mut self) -> Self {
+        self.port = 0;
+        self
+    }
+
+    /// Sets the p2p and discovery ports to zero, allowing the OD to assign a random unused port
+    /// when network components bind to sockets.
+    pub fn with_unused_ports(mut self) -> Self {
+        self = self.with_unused_p2p_port();
+        self.discovery = self.discovery.with_unused_discovery_port();
+        self
+    }
 }
 
 impl Default for NetworkArgs {
@@ -182,6 +197,13 @@ impl DiscoveryArgs {
             network_config_builder = network_config_builder.disable_discv4_discovery();
         }
         network_config_builder
+    }
+
+    /// Set the discovery port to zero, to allow the OS to assign a random unused port when
+    /// discovery binds to the socket.
+    pub fn with_unused_discovery_port(mut self) -> Self {
+        self.port = 0;
+        self
     }
 }
