+++ examples/cli-extension-event-hooks/src/main.rs
@@ -16,10 +16,12 @@
 //! once the node has been started.
 
 use clap::Parser;
-use reth::cli::{
-    components::RethNodeComponents,
-    ext::{NoArgsCliExt, RethNodeCommandConfig},
-    Cli,
+use reth::{
+    cli::Cli,
+    node_core::cli::{
+        components::RethNodeComponents,
+        ext::{NoArgsCliExt, RethNodeCommandConfig},
+    },
 };
 
 fn main() {
